+++ FBSDKCoreKit/FBSDKCoreKit/FBSDKCoreKit.h
@@ -42,7 +42,6 @@
  #import <FBSDKCoreKit/FBSDKLocation.h>
  #import <FBSDKCoreKit/FBSDKLoggingBehavior.h>
  #import <FBSDKCoreKit/FBSDKSettings.h>
- #import <FBSDKCoreKit/FBSDKTestUsersManager.h>
  #import <FBSDKCoreKit/FBSDKUserAgeRange.h>
  #import <FBSDKCoreKit/FBSDKUtility.h>
 
@@ -52,8 +51,6 @@
   #import <FBSDKCoreKit/FBSDKAppLinkResolver.h>
   #import <FBSDKCoreKit/FBSDKAppLinkResolverRequestBuilder.h>
   #import <FBSDKCoreKit/FBSDKAppLinkResolving.h>
-  #import <FBSDKCoreKit/FBSDKAppLinkReturnToRefererController.h>
-  #import <FBSDKCoreKit/FBSDKAppLinkReturnToRefererView.h>
   #import <FBSDKCoreKit/FBSDKAppLinkTarget.h>
   #import <FBSDKCoreKit/FBSDKAppLinkUtility.h>
   #import <FBSDKCoreKit/FBSDKBridgeAPI.h>
@@ -99,7 +96,6 @@
  #import "FBSDKLoggingBehavior.h"
  #import "FBSDKRandom.h"
  #import "FBSDKSettings.h"
- #import "FBSDKTestUsersManager.h"
  #import "FBSDKUserAgeRange.h"
  #import "FBSDKUtility.h"
 
@@ -109,8 +105,6 @@
   #import "FBSDKAppLinkResolver.h"
   #import "FBSDKAppLinkResolverRequestBuilder.h"
   #import "FBSDKAppLinkResolving.h"
-  #import "FBSDKAppLinkReturnToRefererController.h"
-  #import "FBSDKAppLinkReturnToRefererView.h"
   #import "FBSDKAppLinkTarget.h"
   #import "FBSDKAppLinkUtility.h"
   #import "FBSDKBridgeAPI.h"