+++ FBSDKCoreKit/FBSDKCoreKit/GraphAPI/FBSDKGraphErrorRecoveryProcessor.h
@@ -96,12 +96,6 @@ NS_SWIFT_UNAVAILABLE("")
 @interface FBSDKGraphErrorRecoveryProcessor : NSObject
 
 /**
-  Gets the delegate for the current error being processed.
- */
-@property (nonatomic, weak, readonly, nullable) id<FBSDKGraphErrorRecoveryProcessorDelegate>delegate
-DEPRECATED_MSG_ATTRIBUTE("FBSDKGraphErrorRecoveryProcessor's delegate will be removed in the next major version release.");
-
-/**
   Attempts to process the error, return YES if the error can be processed.
  @param error the error to process.
  @param request the related request that may be reissued.
@@ -110,14 +104,6 @@ DEPRECATED_MSG_ATTRIBUTE("FBSDKGraphErrorRecoveryProcessor's delegate will be re
 - (BOOL)processError:(NSError *)error
              request:(id<FBSDKGraphRequest>)request
             delegate:(nullable id<FBSDKGraphErrorRecoveryProcessorDelegate>)delegate;
-
-/**
-  The callback for FBSDKErrorRecoveryAttempting
- @param didRecover if the recovery succeeded
- @param contextInfo unused
- */
-- (void)didPresentErrorWithRecovery:(BOOL)didRecover contextInfo:(nullable void *)contextInfo
-DEPRECATED_MSG_ATTRIBUTE("didPresentErrorWithRecovery:contextInfo: will be removed in the next major version release.");
 
 @end