+++ FBSDKShareKit/FBSDKShareKit/FBSDKSendButton.m
@@ -36,10 +36,7 @@ FBSDKAppEventName FBSDKAppEventNameFBSDKSendButtonDidTap = @"fb_send_button_did_
 @interface FBSDKSendButton () <FBSDKButtonImpressionTracking>
 @end
 
- #pragma clang diagnostic push
- #pragma clang diagnostic ignored "-Wdeprecated-implementations"
 @implementation FBSDKSendButton
- #pragma clang diagnostic pop
 {
   FBSDKMessageDialog *_dialog;
 }